framework: Report keys properly on Get error messages
There's currently an off-by-one error on the error reporting with keys.
Example: A get or call error on import.foo.bar("baz") would only report
"foo", and import.foo("bar") would just return a blank string ("").

This corrects that. It also shifts the tests over to allow the
validation of the error message as well, and adds a little bit more
coverage.