testing: Circumvent GOPATH when modules are in use
Import path inference currently relies on GOPATH being set. This breaks
when modules are in use and the working repository is not in the GOPATH.

This update supports modules through the use of "go list", first by
using "go list -m" to see if modules are enabled, and then by using a
bare "go list" if they are to detect the import path of the requested
package. In this setup, the ImportPath supplied to TestImportCase
dictates the working directory for the second "go list".