Add license section to README.md
This adds a license part to the end of the README.

The MPL default LICENSE file itself contains no template for adding
attribution, hence I usually add it to the bottom of the READMEs for
the projects that I work on.

Adding the license notice is legal in this way as per advice on
Mozilla's site. This is in lieu of headers existing in every file.

References:
MPL license FAQ, Q22: "Does MPL 2.0 require that the MPL 2.0 license
notice header be included in every file?"
  https://www.mozilla.org/en-US/MPL/2.0/FAQ/

Mozilla license headers:
  https://www.mozilla.org/en-US/MPL/headers/