sdk: allow for multiple function invocations in a single RPC call
This adds support for multiple function calls to be made in a single
RPC call to the plugin.

There are multiple components to this:

* sdk.GetReq has now been updated with a new key subtype, GetKey.
This allows arguments to be associated with a specific key rather
than the request as a whole, which was previously just appended to
the end.

* Logic in framework.Import has been adjusted to account for these
keys, specifically Import.Get has been updated. This was an
incremental update to support the multiple keys and there have not
been any serious changes otherwise.

* The protocol has also been adjusted accordingly. We have also
completely removed the "call" field from the Get_Request message, as
it mainly served as a performance optimization to keep from iterating
over keys and arguments unless absolutely necessary, and these
changes here mean that this is now unavoidable.

As a minor consequence of ensuring that everything is working, we
have also fixed a broken test in the rpc package, the source of which
was an improper assertion on an integer argument value in the mock
import (the Go value bridge converts ints to int64s explicitly).