sdk: add support for receiver data and framework.New
This adds support for:

* Receiver data in sdk.GetReq, and sdk.GetResponse, in the form of
the "Context" field. This will be passed from the runtime in the
event of method calls on data returned from an import during earlier
calls. "Callable" in the response also indicates to the runtime
whether or not a return value can receive methods in future call.

* The introduction of the framework.New interface. When present on an
import root, this will allow for instantiation of any namespace that
the import author wants, depending on the structure of the data
passed along in the receiver.

* The appropriate handling for framework.New, and the protocol
modifications for the new fields.

Co-Authored-By: Matthew Hooker <mwhooker@gmail.com>