framework: ensure null return on deep map get with nil value
A consequence of the multi-key function call work that has been
added in #23 is that any key past a function call will now be called
after the function call.

Example: json.unmarshal("{\"foo\": null}").foo, which would normally
return the map {"foo": null}, now will fully traverse the map,
returning the underlying null.

Tests that we have for the json import in the core runtime have
revealed an issue in the way we deeply traverse raw maps (not
namespaces) in that if the final value is nil, we treat it as
undefined (like it was the result of a Get call), even though our map
value conversion will convert nils found in map return values to zero
values (so nil for a map[string]interface{}).

This commit fixes that logic so that things function as expected
again.