proto: add back the call boolean
This adds back the call boolean that was previously at the top level of
the Get_Request message, now placed in the individual keys of the
request.

This was needed because the protobuf serialization/de-serialization
process will not restore a non-nil empty slice to its correct state
over the wire - it remains nil. As such, the boolean stands in to
determine whether or not we need to initialize an empty slice if the key
is a call to a niladic function.