Add more docs and makefile
This adds the following:

* A Makefile with test, tools, and generate targets.
* A tools.go file so we can version the specific tools we need to
version (currently protobuf and mockery).
* More docs!
* Finally, some updates to the mockery generation for the main package
to ensure that mockery does not bail when generating mock_Import.go
(mock_Import_Closer.go is not generated, and if it is not rotated out
before mockery is run, errors are given and the mock does not generate).

This should working with the repository much easier.