Add Sentinel core installation steps to Makefile
This is required for tests to properly run in Circle.

Ironically, this is currently not possible as we are in the middle of
upgrading the protocol. As such, we are now filtering out the tests in
the "testing" package for now, and will re-add them after the release of
Sentinel 0.11.0.