Makefile: use Sentinel v0.11.0 and re-active tests for "testing"
Due to the protocol flux that was a part of the recent update to the SDK
and runtime, it was not possible to test the acceptance testing
framework in the "testing" package until Sentinel v0.11.0 was released.

Now that this has been done, we can start testing this stuff again.