Circle: fix Sentinel test binary caching for CI
This fixes the Sentinel test binary caching for Circle.

The current setup was actually completely ineffective as using a
"restore_cache" clause with two keys only restores one key at a time.
This means if there was a go module cache, Sentinel would be skipped.

Additionally, we've observed in other repositories that the Sentinel
cache restore was failing altogether with permission issues. As such,
this also moves the Sentinel binary to a place where it can be restored
with non-root permissions (we use the "go install" binary path here).