Use .exe extension on built imports
Previous the testing framework would build the import without a .exe
extension on Windows. This meant that sentinel could not craft a
configuration which would make it execute the import file. In turn
this meant that the SDK could not be used to test imports on Windows
systems which is one of the use cases for this SDK

This commit updates the import builder to add the .exe extension
on Windows systems.